light_requirements.txt
tqdm
psycopg2-binary
boto3
scikit-learn
pytest
moto
sqlalchemy
Flask
python-dotenv
torch
spacy
transformers